requirements.txt
kivy
kivymd
pillow
